testify.suite for integration tests
